requirements.txt
amqp==5.3.1
asgiref==3.9.1
astroid==3.3.11
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
dill==0.4.0
Django==5.2.4
django-cors-headers==4.7.0
django-filter==25.1
django-phonenumber-field==8.1.0
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
ecdsa==0.19.1
inflection==0.5.1
isort==6.0.1
kombu==5.5.4
MarkupSafe==3.0.2
mccabe==0.7.0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.10.1
pylint==3.3.8
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
types-PyYAML==6.0.12.20250809
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
